Add Hive type adapter for BookEntity class
This commit adds a Hive type adapter for the BookEntity class, allowing it to be serialized and deserialized using Hive. The BookEntity class now includes HiveField annotations for each of its properties, and a BookEntityAdapter class has been generated using the TypeAdapterGenerator. This will enable the BookEntity class to be stored in a Hive box.